requirements.txt
streamlit
pyyaml
pydicom
pandas
streamlit_authenticator